restructured team tables and fixed filtering issue in teamroleslist
