+++ src/pages/projects/[projectId]/teams/index.tsx
@@ -5,27 +5,25 @@ import Link from "next/link"
 import { useQuery } from "@blitzjs/rpc"
 import { useParam } from "@blitzjs/next"
 import Layout from "src/core/layouts/Layout"
-import {
-  TeamInformation,
-  projectMemberTeamTableColumns,
-  pmTeamTableColumns,
-} from "src/teams/components/TeamTable"
+import { ContributorTeamColumns } from "src/teams/tables/columns/ContributorTeamColumns"
+import { PmTeamColumns } from "src/teams/tables/columns/PmTeamColumns"
 import Table from "src/core/components/Table"
 import { useMemberPrivileges } from "src/projectprivileges/components/MemberPrivilegesContext"
 import { MemberPrivileges, ProjectMember, User } from "@prisma/client"
 import getProjectMembers from "src/projectmembers/queries/getProjectMembers"
 import { useCurrentUser } from "src/users/hooks/useCurrentUser"
+import { processTeam } from "src/teams/tables/processing/processTeam"
 
 interface AllTeamListProps {
   privilege: MemberPrivileges
+  projectId: number | undefined
 }
 
 export type ProjectMemberWithUsers = ProjectMember & {
   users: User[]
 }
 
-export const AllTeamList = ({ privilege }: AllTeamListProps) => {
-  const projectId = useParam("projectId", "number")
+export const AllTeamList = ({ privilege, projectId }: AllTeamListProps) => {
   const currentUser = useCurrentUser()
 
   // use this to get teams
@@ -45,28 +43,19 @@ export const AllTeamList = ({ privilege }: AllTeamListProps) => {
 
   // Filter teams if the privilege is CONTRIBUTOR
   // Now explicitly type the teams to avoid the error
-  const filteredTeams =
-    privilege === MemberPrivileges.CONTRIBUTOR
-      ? projectMembers.filter((projectMember: ProjectMemberWithUsers) =>
-          projectMember.users.some((user) => user.id === currentUser?.id)
-        )
-      : projectMembers
-
-  let teamInformation: TeamInformation[] = filteredTeams.map((team) => {
-    let t: TeamInformation = {
-      name: team.name ?? "Unknown",
-      id: team.id,
-      projectId: projectId,
-    }
-    return t
-  })
+  const teamData = processTeam(
+    projectMembers as ProjectMemberWithUsers[],
+    privilege,
+    currentUser?.id,
+    projectId!
+  )
 
   const tableColumns =
-    privilege === MemberPrivileges.CONTRIBUTOR ? projectMemberTeamTableColumns : pmTeamTableColumns
+    privilege === MemberPrivileges.CONTRIBUTOR ? ContributorTeamColumns : PmTeamColumns
 
   return (
     <div>
-      <Table columns={tableColumns} data={teamInformation} addPagination={true} />
+      <Table columns={tableColumns} data={teamData} addPagination={true} />
     </div>
   )
 }
@@ -84,12 +73,9 @@ const TeamsPage = () => {
 
       <main className="flex flex-col mt-2 mx-auto w-full max-w-7xl">
         <h1 className="flex justify-center mb-2 text-3xl">Teams</h1>
-
-        {
-          <Suspense fallback={<div>Loading...</div>}>
-            <AllTeamList privilege={privilege!} />
-          </Suspense>
-        }
+        <Suspense fallback={<div>Loading...</div>}>
+          <AllTeamList privilege={privilege!} projectId={projectId} />
+        </Suspense>
         {privilege === MemberPrivileges.PROJECT_MANAGER && (
           <div>
             <Link
