+++ src/tasklogs/components/TeamTaskLogTable.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { ColumnDef, createColumnHelper } from "@tanstack/react-table"
 import { TaskLogToggleModal } from "./TaskLogToggleModal"
-import { ShowTeamModal } from "./ShowTeamModal"
+import { ShowTeamModal } from "../../teams/components/ShowTeamModal"
 import { ProcessedTeamTaskLog } from "../utils/processTaskLogs"
 import { TaskLogSchemaModal } from "./TaskLogSchemaModal"
