+++ src/teams/components/ShowTeamModal.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react"
 import Modal from "src/core/components/Modal"
 import { Tooltip } from "react-tooltip"
-import { TeamMembers, teamMembersTableColumns } from "src/teams/components/TeamMembersTable"
+import { TeamMembers, TeamMembersColumn } from "src/teams/tables/columns/TeamMembersColumn"
 import Table from "src/core/components/Table"
 
 export const ShowTeamModal = ({ projectMember }) => {
@@ -38,7 +38,7 @@ export const ShowTeamModal = ({ projectMember }) => {
       <Modal open={openModal} size="w-7/6 max-w-1xl">
         <div className="flex flex-col justify-start mt-4">
           <h3>Team members</h3>
-          <Table columns={teamMembersTableColumns} data={teamMembers} addPagination={true} />
+          <Table columns={TeamMembersColumn} data={teamMembers} addPagination={true} />
         </div>
         <div className="modal-action flex justify-end mt-4">
           <button
