+++ src/teams/components/TeamTaskListDone.tsx
@@ -1,12 +1,13 @@
 import Table from "src/core/components/Table"
 import { useTeamTaskListDone } from "../hooks/useTeamTaskListDone"
+import { TeamTaskListDoneColumns } from "../tables/columns/TeamTaskListDoneColumns"
 
 export const TeamTaskListDone = ({ teamId }) => {
-  const { data, columns } = useTeamTaskListDone(teamId)
+  const { teamTaskListDoneData } = useTeamTaskListDone(teamId)
 
   return (
     <div>
-      <Table columns={columns} data={data} addPagination={true} />
+      <Table columns={TeamTaskListDoneColumns} data={teamTaskListDoneData} addPagination={true} />
     </div>
   )
 }
