+++ src/teams/tables/columns/PmTeamColumns.tsx
@@ -2,17 +2,13 @@ import React from "react"
 import { createColumnHelper } from "@tanstack/react-table"
 import Link from "next/link"
 import { Routes } from "@blitzjs/next"
+import { TeamData } from "../processing/processTeam"
 
-export type TeamInformation = {
-  name: string
-  id: number
-  projectId?: number
-}
 // Column helper
-const columnHelper = createColumnHelper<TeamInformation>()
+const columnHelper = createColumnHelper<TeamData>()
 
 // ColumnDefs
-export const pmTeamTableColumns = [
+export const PmTeamColumns = [
   columnHelper.accessor("name", {
     cell: (info) => <span>{info.getValue()}</span>,
     header: "Team Name",
@@ -51,32 +47,6 @@ export const pmTeamTableColumns = [
           })}
         >
           Edit Team
-        </Link>
-      </div>
-    ),
-  }),
-]
-
-export const projectMemberTeamTableColumns = [
-  columnHelper.accessor("name", {
-    cell: (info) => <span>{info.getValue()}</span>,
-    header: "Team Name",
-  }),
-  columnHelper.accessor("id", {
-    id: "view",
-    header: "View",
-    enableColumnFilter: false,
-    enableSorting: false,
-    cell: (info) => (
-      <div className="">
-        <Link
-          className="btn btn-primary"
-          href={Routes.ShowTeamPage({
-            projectId: info.row.original.projectId!,
-            teamId: info.getValue(),
-          })}
-        >
-          See Contributions
         </Link>
       </div>
     ),
